requirements.txt
# Web scraping and browser automation
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Data processing
pandas==2.1.4
pydantic==2.5.2

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
APScheduler==3.10.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0

# Monitoring and logging
loguru==0.7.2
tqdm==4.66.1